그래픽 엔진 일부 수정
종횡비를 이상하게 가져오는 버그 수정

+++ src/ThruthGameEngine/GraphicsManager.cpp
@@ -31,7 +31,7 @@ void Truth::GraphicsManager::Initalize(HWND _hwnd, uint32 _wight, uint32 _height
 	m_renderScene = m_renderer->CreateRenderScene();
 	m_renderer->SetRenderScene(m_renderScene);
 
-	m_aspect = static_cast<float>(_wight / _height);
+	m_aspect = static_cast<float>(_wight) / static_cast<float>(_height);
 }
 
 void Truth::GraphicsManager::Finalize()
